requirements.txt
requests==2.20.0
python-dotenv==0.19.2